+++ tools/migrate_discord_students.py
@@ -51,7 +51,7 @@ def connect_to_mongo_db(connection_url: str) -> MongoClient:
     try:
         return MongoClient(connection_url)
     except ConnectionError as e:
-        logger.fatal(f"Impossible de se connecter à MongoDB. {e}")
+        logger.fatal(f"connect_to_mongo_db - Impossible de se connecter à MongoDB. {e}")
         exit(-1)
